Made a new solver that seams to be alot faster. This one uses a hashtable. Still not the most optimal solution, and I don't know why yet.

+++ X-code/Lego-Project/SolverNode.hpp
@@ -30,11 +30,11 @@ struct SolverNode{
     std::string graphString_BeforeDiamondMove;
     std::string graphString_AfterDiamondMove;
     
-    long ID;
-    long prevID;
+    int ID;
+    int prevID;
     
     // Distance the robot has traveled to this move:
-    long distanceTraveled;
+    int distanceTraveled;
     
     // Depth in the tree:
     int depthInTree;
